requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
streamlit==1.28.1
plotly==5.17.0 